patentage

+++ Solution/JeanMarcGuaySiteWeb/JeanMarcGuaySiteWeb/Admin/DocumentManagement.aspx.cs
@@ -36,6 +36,7 @@ namespace JeanMarcGuaySiteWeb.Admin
             //Feed les catégories 
             if (!Page.IsPostBack)
             {
+                publicationTable.Rows.Clear();
                 CategoryFactory cf = new CategoryFactory(cnnStr);
                 Category[] categories = cf.GetAll();
 
@@ -58,7 +59,7 @@ namespace JeanMarcGuaySiteWeb.Admin
         protected void SelectedIndexChanged(object sender, EventArgs e)
         {
             string categorie = DdlCategories.SelectedValue;
-
+            publicationTable.Rows.Clear();
             if (categorie == "Tous")
             {
                 Publication[] publications = pf.GetAll();
@@ -90,7 +91,7 @@ namespace JeanMarcGuaySiteWeb.Admin
                 cellTelecharger.Controls.Add(buttonDownload);
 
                 Button buttonSupprimer = new Button();
-                buttonSupprimer.Attributes.Add("class", "btn btn-warning");
+                buttonSupprimer.Attributes.Add("class", "btn btn-warning btnSuppr");
                 buttonSupprimer.Text = "Supprimer";
                 buttonSupprimer.Attributes.Add("data-id", publication.publicationId.ToString());
                 buttonSupprimer.Click += new EventHandler(btn_Supprimer_Click);
@@ -117,19 +118,23 @@ namespace JeanMarcGuaySiteWeb.Admin
         }
         protected void btn_Supprimer_Click(object sender, EventArgs e)
         {
-            Button button = (Button)sender;
-            int id = Convert.ToInt32(button.Attributes["data-id"]);
-            Publication publication = pf.Get(id);
-            pf.delete(id);
-            try
+            string confirmValue = Request.Form["confirm_delete"];
+            if (confirmValue == "Oui")
             {
-                string FileToDelete = Server.MapPath(publication.url);
-                File.Delete(FileToDelete);
-            }
-            finally
-            {
-                Response.Redirect(Request.RawUrl);
-            }      
+                Button button = (Button)sender;
+                int id = Convert.ToInt32(button.Attributes["data-id"]);
+                Publication publication = pf.Get(id);
+                pf.delete(id);
+                try
+                {
+                    string FileToDelete = Server.MapPath(publication.url);
+                    File.Delete(FileToDelete);
+                }
+                finally
+                {
+                    Response.Redirect(Request.RawUrl);
+                }
+            } 
         }
